requirements.txt
flask
flask-jwt-extended
Flask-SQLAlchemy
Flask-Migrate
flask-restful
pytest-flask
psycopg2-binary
